Atomizes beta banner
