+++ cfgov/unprocessed/css/main.less
@@ -73,6 +73,7 @@
 @import (less) "molecules/featured-menu-content.less";
 @import (less) "molecules/form-field-with-button.less";
 @import (less) "molecules/full-width-text.less";
+@import (less) "molecules/global-banner.less";
 @import (less) "molecules/global-eyebrow.less";
 @import (less) "molecules/global-header-cta.less";
 @import (less) "molecules/global-search.less";
@@ -118,7 +119,6 @@
 /* Page-specific styles
    ========================================================================== */
 
-@import (less) "pages/beta.less";
 @import (less) "pages/business.less";
 @import (less) "pages/event.less";
 @import (less) "pages/external-site.less";
