+++ cfgov/unprocessed/css/organisms/header.less
@@ -80,6 +80,7 @@
     } );
 
     &_content {
+        position: relative;
 
         > .wrapper {
             > .m-global-search {
